+++ hal/include/HAL/PWM.h
@@ -18,9 +18,9 @@ void HAL_FreePWMPort(HAL_DigitalHandle pwm_port_handle, int32_t* status);
 
 bool HAL_CheckPWMChannel(int32_t pin);
 
-void HAL_SetPWMConfig(HAL_DigitalHandle pwm_port_handle, float maxPwm,
-                      float deadbandMaxPwm, float centerPwm,
-                      float deadbandMinPwm, float minPwm, int32_t* status);
+void HAL_SetPWMConfig(HAL_DigitalHandle pwm_port_handle, double maxPwm,
+                      double deadbandMaxPwm, double centerPwm,
+                      double deadbandMinPwm, double minPwm, int32_t* status);
 void HAL_SetPWMConfigRaw(HAL_DigitalHandle pwm_port_handle, int32_t maxPwm,
                          int32_t deadbandMaxPwm, int32_t centerPwm,
                          int32_t deadbandMinPwm, int32_t minPwm,
@@ -35,14 +35,14 @@ HAL_Bool HAL_GetPWMEliminateDeadband(HAL_DigitalHandle pwm_port_handle,
                                      int32_t* status);
 void HAL_SetPWMRaw(HAL_DigitalHandle pwm_port_handle, int32_t value,
                    int32_t* status);
-void HAL_SetPWMSpeed(HAL_DigitalHandle pwm_port_handle, float speed,
+void HAL_SetPWMSpeed(HAL_DigitalHandle pwm_port_handle, double speed,
                      int32_t* status);
-void HAL_SetPWMPosition(HAL_DigitalHandle pwm_port_handle, float position,
+void HAL_SetPWMPosition(HAL_DigitalHandle pwm_port_handle, double position,
                         int32_t* status);
 void HAL_SetPWMDisabled(HAL_DigitalHandle pwm_port_handle, int32_t* status);
 int32_t HAL_GetPWMRaw(HAL_DigitalHandle pwm_port_handle, int32_t* status);
-float HAL_GetPWMSpeed(HAL_DigitalHandle pwm_port_handle, int32_t* status);
-float HAL_GetPWMPosition(HAL_DigitalHandle pwm_port_handle, int32_t* status);
+double HAL_GetPWMSpeed(HAL_DigitalHandle pwm_port_handle, int32_t* status);
+double HAL_GetPWMPosition(HAL_DigitalHandle pwm_port_handle, int32_t* status);
 void HAL_LatchPWMZero(HAL_DigitalHandle pwm_port_handle, int32_t* status);
 void HAL_SetPWMPeriodScale(HAL_DigitalHandle pwm_port_handle,
                            int32_t squelchMask, int32_t* status);
